+++ CHANGELOG.md
@@ -2,7 +2,17 @@
 
 All notable changes to this project will be documented in this file.
 
-## 1.0.3
+## 1.1.0
+
+* Spring-Security 5 integration libraries. Added AudienceValidator
+* Spring-Security 5 Support for basic authentication
+
+## 1.1.0.RC1
+
+* Initial version including spring-security 5 integration libraries
+
+
+## 1.0.0
 
 * Initial version of the api for SAP Java Buildpack